Fix error: multiple assignments to variable 'R'
The variable 'R' is used within the residual to represent both the
universal gas constant (k_B*N_A) and the average particle radius.  The
assignment inside the function is for the latter, so any use of `R`
intended to mean the gas constant results in an error.

At t=0, the particle radius is zero, and so division by R returns `nan`
in all the resdiuals.  Surprisingly, the integrator does not quit or
return an error, but rather chugs along happily, forever :(

Fixed by disambiguating the two.  `Rbar` represents the gas constant.

I also changed the solver method: "BDF" gives a smoother profile
that lets us integrate out to longer times, where we observe a steady state.

Yay - looks great!